Merge develop to master

See merge request QuentG/atoute-back!8

+++ src/Controller/OfferController.php
@@ -4,13 +4,14 @@ namespace App\Controller;
 
 use App\Entity\Offer;
 use App\Form\OfferType;
+use App\Security\Voter\OfferVoter;
 use App\Repository\OfferRepository;
-use App\Service\AuthService;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\Form\FormFactoryInterface;
-use Symfony\Component\HttpFoundation\JsonResponse;
+use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Form\FormFactoryInterface;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Serializer\SerializerInterface;
 
 /**
@@ -22,20 +23,16 @@ final class OfferController extends BaseController
     private EntityManagerInterface $entityManager;
     private FormFactoryInterface $formFactory;
     private OfferRepository $offerRepository;
-    private AuthService $authService;
 
     public function __construct(
         SerializerInterface $serializer,
         EntityManagerInterface $entityManager,
-        FormFactoryInterface $formFactory,
-        OfferRepository $offerRepository,
-        AuthService $authService
+        FormFactoryInterface $formFactory
     ) {
         $this->serializer = $serializer;
         $this->entityManager = $entityManager;
         $this->formFactory = $formFactory;
-        $this->offerRepository = $offerRepository;
-        $this->authService = $authService;
+        $this->offerRepository = $entityManager->getRepository(Offer::class);
     }
 
     /**
@@ -44,6 +41,10 @@ final class OfferController extends BaseController
     public function create(Request $request): JsonResponse
     {
         $data = $this->testJson($request);
+
+        if (!$this->isGranted(OfferVoter::CREATE)) {
+            return $this->respondWithError('only_companies_can_create_offer');
+        }
 
         $form = $this->formFactory->create(OfferType::class);
         $form->submit($data);
@@ -71,11 +72,46 @@ final class OfferController extends BaseController
     }
 
     /**
+     * @Route("", name="all", methods={"GET"})
+     */
+    public function all(Request $request, PaginatorInterface $paginator): JsonResponse
+    {
+        $limit = (int) $request->query->get('l') !== 0 ? (int) $request->query->get('l') : null;
+        $offerPerPage = (int) $request->query->get('n') !== 0 ? (int) $request->query->get('n') : null;
+
+        // Filter values
+        $type = (string) $request->query->get('type') !== "" ? (string) $request->query->get('type') : null;
+        $activity = (string) $request->query->get('activity') !== "" ? (string) $request->query->get('activity') : null;
+        // Filters array
+        $filters = array();
+        $filters["type"] = $type;
+        $filters["activity"] = $activity;
+
+        $pagination = $paginator->paginate(
+            $this->offerRepository->getPublishQuery($limit, $filters), // Requête contenant les données à paginer (ici nos articles)
+            $request->query->getInt('page', 1), // Numéro de la page en cours, passé dans l'URL, 1 si aucune page
+            $offerPerPage// Nombre de résultats par page
+        );
+
+        $json = $this->serializer->serialize(
+            $pagination,
+            'json',
+            ['groups' => 'offer_read']
+        );
+
+        return $this->respond('offers_infos', json_decode($json));
+    }
+
+    /**
      * @Route("/{id}", name="index", methods={"GET"})
      */
     public function index(int $id): JsonResponse
     {
-        $offer = $this->getAndVerifyOffer($id, false);
+        $offer = $this->getAndVerifyOffer($id);
+
+        if (!$offer instanceof Offer) {
+            return $this->respondWithError($offer);
+        }
 
         $json = $this->serializer->serialize(
             $offer,
@@ -95,6 +131,10 @@ final class OfferController extends BaseController
 
         if (!$offer instanceof Offer) {
             return $this->respondWithError($offer);
+        }
+
+        if (!$this->isGranted(OfferVoter::EDIT, $offer)) {
+            return $this->respondWithError('only_companies_can_create_offer');
         }
 
         $data = $this->testJson($request);
@@ -126,6 +166,10 @@ final class OfferController extends BaseController
             return $this->respondWithError($offer);
         }
 
+        if (!$this->isGranted(OfferVoter::EDIT, $offer)) {
+            return $this->respondWithError('not_offer_owner');
+        }
+
         $this->entityManager->remove($offer);
         $this->entityManager->flush();
 
@@ -133,16 +177,35 @@ final class OfferController extends BaseController
     }
 
     /**
+     * @Route("/{id}/applications", name="applications", methods={"GET"})
+     */
+    public function applications(int $id): JsonResponse
+    {
+        $offer = $this->getAndVerifyOffer($id);
+
+        if (!$offer instanceof Offer) {
+            return $this->respondWithError($offer);
+        }
+
+        if (!$this->isGranted(OfferVoter::EDIT, $offer)) {
+            return $this->respondWithError('not_offer_owner');
+        }
+
+        $applications = [];
+        foreach ($offer->getApplicationsToBeProcessed() as $application) {
+            $applications[] = json_decode($this->serializer->serialize($application, 'json', ['groups' => 'application_offer_read']));
+        }
+
+        return $this->respond('', $applications);
+    }
+
+    /**
      * @return Offer|string
      */
-    private function getAndVerifyOffer(int $id, bool $verifyOwner = true)
+    private function getAndVerifyOffer(int $id)
     {
         if (!$offer = $this->offerRepository->find($id)) {
             return 'offer_not_found';
-        }
-
-        if ($verifyOwner && !$offer->isOwner($this->authService->getUser())) {
-            return 'bad_offer_owner';
         }
 
         return $offer;
